profile-backend/Backend/requirements.txt
Django==3.1.6
Jinja2==2.11.3
MarkupSafe==1.1.1
PyYAML==5.4.1
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
docker==4.4.1
docker-compose==1.28.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
isort==5.7.0
itypes==1.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql==0.0.2
mysql-connector-python==8.0.22
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==20.9
pep8==1.7.1
pip==21.0.1
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pytz==2021.1
pywin32==227
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
setuptools==49.2.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
websocket-client==0.57.0
wrapt==1.12.1
django-heroku